refactor: QuestionGroupPanel を 3 コンポーネントに分割
